moved benchmark-like tests to benchmark folder inside tests and renamed test methods for easier understanding
