+++ tests/benchmarks/test_challenge_player.py
@@ -3,20 +3,20 @@ import pytest
 from pyschieber.player.challenge_player.challenge_player import ChallengePlayer
 from pyschieber.player.greedy_player.greedy_player import GreedyPlayer
 from pyschieber.player.random_player import RandomPlayer
-from tests.example.statistical_helper import run_statistics
+from tests.benchmarks.statistical_helper import run_statistics
 
 
 @pytest.mark.statistical
-def test_greedy():
-    players = [GreedyPlayer(name='Greedy1'), RandomPlayer(name='Track1'), GreedyPlayer(name='Greedy2'),
-               RandomPlayer(name='Track2')]
+def test_against_random():
+    players = [ChallengePlayer(name='Challenge1'), RandomPlayer(name='Random1'), ChallengePlayer(name='Challenge2'),
+               RandomPlayer(name='Random2')]
 
     run_statistics(players=players)
 
 
 @pytest.mark.statistical
-def test_challenge():
-    players = [ChallengePlayer(name='Trick1'), GreedyPlayer(name='Greedy1'), ChallengePlayer(name='Trick2'),
+def test_against_greedy():
+    players = [ChallengePlayer(name='Challenge1'), GreedyPlayer(name='Greedy1'), ChallengePlayer(name='Challenge2'),
                GreedyPlayer(name='Greedy2')]
     
     run_statistics(players=players)
